--- lib/src/models/countly_presentable_feedback.dart
@@ -1,11 +1,0 @@
-class CountlyPresentableFeedback {
-  CountlyPresentableFeedback(this.widgetId, this.type, this.name);
-
-  final String widgetId;
-  final String type;
-  final String name;
-
-  static CountlyPresentableFeedback fromJson(dynamic json) {
-    return CountlyPresentableFeedback(json['id'], json['type'], json['name']);
-  }
-}
