--- lib/src/models/feedback_widget_response.dart
@@ -1,8 +1,0 @@
-import 'countly_presentable_feedback.dart';
-
-class FeedbackWidgetsResponse {
-  FeedbackWidgetsResponse(this.presentableFeedback, this.error);
-
-  final String? error;
-  final List<CountlyPresentableFeedback> presentableFeedback;
-}
