+++ pubspec.yaml
@@ -12,7 +12,6 @@ dependencies:
   flutter:
     sdk: flutter
   pedantic: ^1.11.0
-  js: ^0.6.3
 
 dev_dependencies:
   flutter_test:
